Connecting to MongoDB:
Inside the connectDB function, the await keyword is used to wait for the connection to the database.
mongoose.connect(process.env.MONGO_URI) is called, where mongoose.connect is a method provided by mongoose to establish a connection to a MongoDB database.
process.env.MONGO_URI is an environment variable that holds the connection string (URI) needed to connect to your MongoDB instance. Using environment variables for sensitive information like database credentials is a common practice to keep your code secure.