Schema Fields:
fullName:
type: String: The fullName field is of type String.
required: true: This field is mandatory, meaning a value must be provided when creating a document.
username:
type: String: The username field is of type String.
required: true: This field is mandatory.
unique: true: This ensures that each value for the username field is unique across all documents in the collection.
password:
type: String: The password field is of type String.
required: true: This field is mandatory.
profilePhoto:
type: String: The profilePhoto field is of type String.
default: "": If no value is provided for this field, it defaults to an empty string.
gender:
type: String: The gender field is of type String.
enum: ["male", "female"]: This limits the values to either "male" or "female".
required: true: This field is mandatory.