Defining Routes Register Route:
router.route("/register").post(register);: Defines a POST route for /register.
When a POST request is made to /register, the register function from the controller is executed. This handles user registration.
Login Route:

router.route("/login").post(login);: Defines a POST route for /login.
When a POST request is made to /login, the login function from the controller is executed. This handles user login.
Logout Route:

router.route("/logout").get(logout);: Defines a GET route for /logout.
When a GET request is made to /logout, the logout function from the controller is executed. This handles user logout.
Get Other Users Route:

router.route("/").get(isAuthenticated, getOtherUsers);: Defines a GET route for /.
When a GET request is made to /, the isAuthenticated middleware is first executed to check if the user is authenticated.
If the user is authenticated, the getOtherUsers function from the controller is executed to get a list of other users.