Route Definitions Route for Sending Messages:
router.route("/send/:id").post(isAuthenticated, sendMessage);:
Defines a route using router.route(). This route handles POST requests (post()) to /send/:id.
:id in the route path is a route parameter that represents the receiverId, which is passed as part of the URL.
isAuthenticated is a middleware function that ensures the user is authenticated before allowing access to the route.
sendMessage is the controller function imported from messageController.js. It handles the logic for sending a message.
When a POST request is made to /send/:id, the isAuthenticated middleware first checks if the user is authenticated. If authenticated, it then calls the sendMessage controller function to handle sending the message.
Route for Getting Messages:

router.route("/:id").get(isAuthenticated, getMessage);:
Defines a route using router.route(). This route handles GET requests (get()) to /:id.
:id in the route path represents the receiverId, which is passed as part of the URL to retrieve messages for a specific conversation.
isAuthenticated is again used as middleware to ensure the user is authenticated before accessing the route.
getMessage is the controller function imported from messageController.js. It handles the logic for retrieving messages.
When a GET request is made to /:id, the isAuthenticated middleware verifies the user's authentication. If authenticated, it then calls the getMessage controller function to retrieve and return messages for the conversation with the specified receiverId.