export default messageSlice.reducer;:
Exports the messageSlice.reducer, which is a generated reducer function that handles actions defined in the reducers property of createSlice.
This reducer function manages updates to the slice's state (initialState) based on dispatched actions.
Summary
Purpose: Defines a Redux slice named "message" with initial state containing a messages field initialized to null.
Reducer: Implements a reducer function (setMessages) to update the messages state field based on dispatched actions.
Export: Makes the action creator (setMessages) and the reducer function (messageSlice.reducer) available for use in connecting components or middleware to the Redux store.