JSX (JSX Markup) Form Container:
The form is wrapped inside a div with classes for styling, including padding, rounded corners, shadow, and background styling.
Form Header:

Displays a header with the text "Signup".
Form Fields:

Full Name:
An input field for entering the full name. It updates the fullName property in the user state on change.
Username:
An input field for entering the username. It updates the username property in the user state on change.
Password:
An input field for entering the password. It updates the password property in the user state on change.
Confirm Password:
An input field for confirming the password. It updates the confirmPassword property in the user state on change.
Gender Selection:

A section with two checkboxes for selecting the gender (Male or Female). The checked property is determined by comparing the gender property in the user state with the respective gender string. The onChange event calls the handleCheckbox function.
Login Link:

A paragraph with a link to the login page for users who already have an account.
Submit Button:

A button to submit the form. It triggers the onSubmitHandler function when clicked.
Styles
Tailwind CSS:
The component uses Tailwind CSS utility classes for styling, such as min-w-96, mx-auto, w-full, p-6, rounded-lg, shadow-md, bg-gray-400, bg-clip-padding, backdrop-filter, backdrop-blur-md, bg-opacity-10, border, border-gray-100, text-3xl, font-bold, text-center, label, label-text, input, input-bordered, h-10, my-4, flex, items-center, mx-2, text-center, my-2, btn, btn-block, btn-sm, mt-2, and border-slate-700.