Imports React and useEffect:
React is the core library for building user interfaces.
useEffect is a hook used for side effects, though it is not utilized in this particular code snippet.
SendInput and Messages:

These are custom components used within MessageContainer for sending and displaying messages.
useSelector and useDispatch from react-redux:

useSelector is used to access specific parts of the Redux store's state.
useDispatch is used to dispatch actions to the Redux store.
setSelectedUser:

An action creator imported from ../redux/userSlice to set the selected user in the Redux store.