Imports React, useEffect, useRef:
React: Core library for building the user interface.
useEffect: A hook for performing side effects in functional components.
useRef: A hook for creating a mutable object that persists across renders and can hold a reference to a DOM element.
useSelector:

A hook from react-redux to access the Redux store's state.